create table(task) and fix some file(web.app,controller,view file)

+++ app/Http/Controllers/TasksController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 
 use App\Task;
+use App\User;
 
 class TasksController extends Controller
 {
@@ -15,11 +16,25 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        // まず空の一次配列を用意
+        $data = [];
+        
+        // ログインしてたらユーザとそのタスクを取得して$dataの配列に格納
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        // 追記：$dataの配列にタスクの個数をカウントして渡す        
+        $data += $this->counts($user);
+
+        //index.blade.phpに取得したuser情報とタスクを渡す
+        //indexページをViewとして表示
+        return view('tasks.index', $data);
     }
 
     /**
@@ -29,6 +44,7 @@ class TasksController extends Controller
      */
     public function create()
     {
+        /* 変更なし */
         $task = new Task;
 
         return view('tasks.create', [
@@ -44,16 +60,17 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
-        //バリデーション
+        // バリデーション
         $this->validate($request, [
             'content' => 'required|max:191',
             'status' => 'required|max:10',
         ]);
 
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        // 追加
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status'  => $request->status,
+        ]);
 
         return redirect('/');
     }
@@ -66,7 +83,7 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        //
+        /* 変更なし */
         $task = Task::find($id);
 
         return view('tasks.show', [
@@ -82,6 +99,7 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
+        /* 変更なし */
         $task = Task::find($id);
 
         return view('tasks.edit', [
@@ -104,6 +122,7 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
+        /* 変更なし */
         $task = Task::find($id);
         $task->content = $request->content;
         $task->status = $request->status;
@@ -118,11 +137,19 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    
     public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        // Taskモデルが所有してるタスクのうち、選んだidのタスクを探す
+        $task = \App\Task::find($id);
 
+        // 選んだid が 編集中のユーザ（user_id) かどうかを検証
+        // 正しければ削除を実行。トップページに戻る。
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+        
         return redirect('/');
     }
+
 }
